backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
loguru==0.7.2
openai==1.3.0
python-dateutil==2.8.2
